Moving Link to 'types' package
